extracted options and assertions into separate files and some refactoring
